# Conflicts:
#	mobile/composeApp/src/androidMain/kotlin/com/yagubogu/YaguBoguApplication.kt
#	mobile/composeApp/src/commonMain/kotlin/com/yagubogu/di/DataSourceModule.kt
#	mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/home/HomeViewModel.kt
#	mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/livetalk/chat/LivetalkChatScreen.kt
#	mobile/composeApp/src/iosMain/kotlin/com/yagubogu/MainViewController.kt
#	mobile/iosApp/iosApp/Info.plist
#	mobile/iosApp/iosApp/iOSApp.swift

+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/data/dto/response/stats/VictoryFairyRankingDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class VictoryFairyRankingDto(
     @SerialName("ranking")
-    val ranking: Int, // 승리요정 랭킹
+    val ranking: Long, // 승리요정 랭킹
     @SerialName("memberId")
     val memberId: Long, // 멤버 ID
     @SerialName("nickname")
